National Life Expectancies data
When we used the linearregression between the fertility and life expectancy (just twovariables problem), we got R2 = 0.6322. While when weconsidered addition two variables for regression, forumlating multi-linear regression problem, the R2 remained almost constant (R2 =0.6307). This means that the addition of these two variables didn’taffect the regression accuracy which means that both of themdon’t have a signficance on the life expectancy. This can beconfirmed when we have a look on the p-value corresponding tothe two regression coefficient associated with the lnHealth andpublic education. So, investment in better monitoring of fertility canlead to be better regression and prediction of life expectancycompared to the rest of studied variables.